requirements.txt
streamlit==1.15.2   # Use the latest stable version if different
pandas==1.5.3       # Use the appropriate version you need
PyMuPDF==1.22.0     # Library for PDF text extraction
python-docx==0.8.11 # Library to handle .docx files
